+++ projects/portfolio/src/client/components/pages/Chat.tsx
@@ -14,18 +14,18 @@ import Markdown from 'react-markdown'
 import remarkGfm from 'remark-gfm'
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter'
 
-import type { AppType } from '../../server/app'
-import { GearIcon } from './svg/GearIcon'
-import { ChatbotIcon } from './svg/ChatbotIcon'
-import { SpinnerIcon } from './svg/SpinnerIcon'
-import { NewChatIcon } from './svg/NewChatIcon'
-import { ChatInput } from './input/ChatInput'
-import { ToggleInput } from './input/ToggleInput'
-import { useResponsive } from './hooks/useResponsive'
-import { ArrowUpIcon } from './svg/ArrowUpIcon'
-import { StopIcon } from './svg/StopIcon'
-import { UploadIcon } from './svg/UploadIcon'
-import { FileImageInput, FileImagePreview } from './input/FileImageInput'
+import type { AppType } from '../../../server/app'
+import { GearIcon } from '../svg/GearIcon'
+import { ChatbotIcon } from '../svg/ChatbotIcon'
+import { SpinnerIcon } from '../svg/SpinnerIcon'
+import { NewChatIcon } from '../svg/NewChatIcon'
+import { ChatInput } from '../input/ChatInput'
+import { ToggleInput } from '../input/ToggleInput'
+import { useResponsive } from '../hooks/useResponsive'
+import { ArrowUpIcon } from '../svg/ArrowUpIcon'
+import { StopIcon } from '../svg/StopIcon'
+import { UploadIcon } from '../svg/UploadIcon'
+import { FileImageInput, FileImagePreview } from '../input/FileImageInput'
 
 const client = hc<AppType>('/')
