+++ projects/portfolio/src/client/components/pages/Profile.tsx
@@ -1,8 +1,8 @@
 import { useState, type FC, type FormEvent } from 'react'
 import { hc } from 'hono/client'
 
-import type { AppType } from '../../server/app'
-import { FileImageInput, FileImagePreview } from './input/FileImageInput'
+import type { AppType } from '../../../server/app'
+import { FileImageInput, FileImagePreview } from '../input/FileImageInput'
 
 const client = hc<AppType>('/')
