+++ projects/portfolio/src/client/routes/about.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import { About } from '../components/About'
+import { About } from '../components/pages/About'
 
 export const Route = createFileRoute('/about')({
   component: RouteComponent,
