+++ projects/portfolio/src/client/routes/chat.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import { Chat } from '../components/Chat'
+import { Chat } from '../components/pages/Chat'
 
 export const Route = createFileRoute('/chat')({
   component: RouteComponent,
