+++ projects/portfolio/src/client/routes/profile.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router'
 
-import { Profile } from '../components/Profile'
+import { Profile } from '../components/pages/Profile'
 
 export const Route = createFileRoute('/profile')({
   component: RouteComponent,
